convert_first_page_to_image/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
pdf2image==1.17.0
pillow==11.0.0
requests==2.32.3
urllib3==2.2.3
